moved move class to its own file removed unneccesary requests import from poke class
